Refactored unit tests (using addresses)
