+++ __test__/ReceiptMerkleTree.spec.ts
@@ -1,429 +1,54 @@
-import test from 'ava'
-
-import { Address } from '@btc-vision/bsi-binary';
-import { MerkleTreeOld, MerkleTreeNew } from './MerkleTree.js';
-import { MAX_HASH, MAX_MINUS_ONE } from './types/ZeroValue.js';
-
-export class ReceiptMerkleTreeOld extends MerkleTreeOld<string, Buffer> {
-    public static TREE_TYPE: [string, string] = ['bytes', 'bytes'];
-
-    constructor() {
-        super(ReceiptMerkleTreeOld.TREE_TYPE);
-    }
-
-    public getProofs(): Map<Address, Map<string, string[]>> {
-        if (!this.tree) {
-            throw new Error('Merkle tree not generated');
-        }
-
-        this.validate();
-
-        const proofs = new Map<Address, Map<string, string[]>>();
-        for (const [address, val] of this.values.entries()) {
-            for (const [key, value] of val.entries()) {
-
-                const transactionBuf = Buffer.from(key, 'hex');
-                const vvv: [Buffer, Buffer] = [transactionBuf, value]
-                const proof: string[] = this.tree.getProof(vvv);
-
-                if (!proof || !proof.length) {
-                    throw new Error(`Proof not found for ${key}`);
-                }
-
-                if (!proofs.has(address)) {
-                    proofs.set(address, new Map());
-                }
-
-                const proofMap = proofs.get(address);
-                if (proofMap) {
-                    proofMap.set(key, proof);
-                }
-            }
-        }
-
-        return proofs;
-    }
-
-    /** We have to replace the value of the given address and key with the new value */
-    public updateValues(address: Address, val: Map<string, Buffer>): void {
-        this.ensureAddress(address);
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        let valueChanged: boolean = false;
-        for (const [key, value] of val.entries()) {
-            const currentValue = map.get(key);
-            if (currentValue && currentValue === value) {
-                continue;
-            }
-
-            map.set(key, value);
-            valueChanged = true;
-        }
-
-        this.valueChanged = valueChanged;
-    }
-
-    public updateValue(contractAddress: string, transactionId: string, result: Uint8Array): void {
-        if (this.frozen) {
-            throw new Error('Merkle tree is frozen, cannot update value');
-        }
-
-        this.ensureAddress(contractAddress);
-
-        const map = this.values.get(contractAddress);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        const currentValue = map.get(transactionId);
-        if (currentValue && currentValue === result) {
-            return;
-        }
-
-        map.set(transactionId, Buffer.from(result));
-        this.valueChanged = true;
-    }
-
-    public getValue(address: string, key: string): Buffer | undefined {
-        if (!this.values.has(address)) {
-            return;
-        }
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        return map.get(key);
-    }
-
-    public getValueWithProofs(address: string, key: string): [Buffer, string[]] | undefined {
-        if (!this.tree) {
-            return;
-        }
-
-        this.validate();
-
-        const keyBuf = Buffer.from(key, 'hex');
-        const value = this.getValue(address, key);
-        if (!value) {
-            return undefined;
-        }
-
-        const proof: string[] = this.tree.getProof([keyBuf, value]);
-        if (!proof || !proof.length) {
-            throw new Error(`Proof not found for ${keyBuf.toString('hex')}`);
-        }
-
-        return [value, proof];
-    }
-
-    public getValuesWithProofs(address: string): Map<string, [Buffer, string[]]> {
-        if (!this.tree) {
-            throw new Error('Merkle tree not generated');
-        }
-
-        this.validate();
-
-        const proofs = new Map<string, [Buffer, string[]]>();
-        if (!this.values.has(address)) {
-            return proofs;
-        }
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        for (const [key, value] of map.entries()) {
-            const keyBuf = Buffer.from(key, 'hex');
-            const proof: string[] = this.tree.getProof([keyBuf, value]);
-
-            if (!proof || !proof.length) {
-                throw new Error(`Proof not found for ${key}`);
-            }
-
-            proofs.set(key, [value, proof]);
-        }
-
-        return proofs;
-    }
-
-    public getEverythingWithProofs(): Map<string, Map<string, [Buffer, string[]]>> | undefined {
-        if (!this.tree) {
-            return;
-        }
-
-        this.validate();
-
-        const proofs = new Map<string, Map<string, [Buffer, string[]]>>();
-        for (const [address] of this.values.entries()) {
-            const map = this.getValuesWithProofs(address);
-
-            proofs.set(address, map);
-        }
-
-        return proofs;
-    }
-
-    public getValues(): [Buffer, Buffer][] {
-        const entries: [Buffer, Buffer][] = [];
-
-        for (const [_address, map] of this.values.entries()) {
-            for (const [key, value] of map.entries()) {
-                const keyBuf = Buffer.from(key, 'hex');
-
-                entries.push([keyBuf, value]);
-            }
-        }
-
-        return entries;
-    }
-
-    protected getDummyValues(): Map<string, Map<string, Buffer>> {
-        const dummyValues = new Map<string, Map<string, Buffer>>();
-        const dummyMap = new Map<string, Buffer>();
-
-        // Ensure minimum tree requirements
-        dummyMap.set(MAX_HASH, Buffer.from([1]));
-        dummyMap.set(MAX_MINUS_ONE, Buffer.from([1]));
-
-        // Add dummy values for the contract
-        dummyValues.set(MAX_MINUS_ONE, dummyMap);
-
-        return dummyValues;
-    }
-
-    private ensureAddress(address: Address): void {
-        if (!this.values.has(address)) {
-            this.values.set(address, new Map());
-        }
-    }
-}
-
-
-
-export class ReceiptMerkleTreeNew extends MerkleTreeNew<string, Buffer> {
-    public static TREE_TYPE: [string, string] = ['bytes', 'bytes'];
-
-    constructor() {
-        super(ReceiptMerkleTreeNew.TREE_TYPE);
-    }
-
-    public getProofs(): Map<Address, Map<string, string[]>> {
-        if (!this.tree) {
-            throw new Error('Merkle tree not generated');
-        }
-
-        const proofs = new Map<Address, Map<string, string[]>>();
-        for (const [address, val] of this.values.entries()) {
-            for (const [key, value] of val.entries()) {
-                const transactionBuf = Buffer.from(key, 'hex');
-                const proof: string[] = this.getProofHashes([transactionBuf, value]);
-
-                if (!proof || !proof.length) {
-                    throw new Error(`Proof not found for ${key}`);
-                }
-
-                if (!proofs.has(address)) {
-                    proofs.set(address, new Map());
-                }
-
-                const proofMap = proofs.get(address);
-                if (proofMap) {
-                    proofMap.set(key, proof);
-                }
-            }
-        }
-
-        return proofs;
-    }
-
-    /** We have to replace the value of the given address and key with the new value */
-    public updateValues(address: Address, val: Map<string, Buffer>): void {
-        this.ensureAddress(address);
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        let valueChanged: boolean = false;
-        for (const [key, value] of val.entries()) {
-            const currentValue = map.get(key);
-            if (currentValue && currentValue === value) {
-                continue;
-            }
-
-            map.set(key, value);
-            valueChanged = true;
-        }
-
-        this.valueChanged = valueChanged;
-    }
-
-    public updateValue(contractAddress: string, transactionId: string, result: Uint8Array): void {
-        if (this.frozen) {
-            throw new Error('Merkle tree is frozen, cannot update value');
-        }
-
-        this.ensureAddress(contractAddress);
-
-        const map = this.values.get(contractAddress);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        const currentValue = map.get(transactionId);
-        if (currentValue && currentValue === result) {
-            return;
-        }
-
-        map.set(transactionId, Buffer.from(result));
-        this.valueChanged = true;
-    }
-
-    public getValue(address: string, key: string): Buffer | undefined {
-        if (!this.values.has(address)) {
-            return;
-        }
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        return map.get(key);
-    }
-
-    public getValueWithProofs(address: string, key: string): [Buffer, string[]] | undefined {
-        if (!this.tree) {
-            return;
-        }
-
-        const keyBuf = Buffer.from(key, 'hex');
-        const value = this.getValue(address, key);
-        if (!value) {
-            return undefined;
-        }
-
-        const proof: string[] = this.getProofHashes([keyBuf, value]);
-        if (!proof || !proof.length) {
-            throw new Error(`Proof not found for ${keyBuf.toString('hex')}`);
-        }
-
-        return [value, proof];
-    }
-
-    public getValuesWithProofs(address: string): Map<string, [Buffer, string[]]> {
-        if (!this.tree) {
-            throw new Error('Merkle tree not generated');
-        }
-
-        const proofs = new Map<string, [Buffer, string[]]>();
-        if (!this.values.has(address)) {
-            return proofs;
-        }
-
-        const map = this.values.get(address);
-        if (!map) {
-            throw new Error('Map not found');
-        }
-
-        for (const [key, value] of map.entries()) {
-            const keyBuf = Buffer.from(key, 'hex');
-            const proof: string[] = this.getProofHashes([keyBuf, value]);
-
-            if (!proof || !proof.length) {
-                throw new Error(`Proof not found for ${key}`);
-            }
-
-            proofs.set(key, [value, proof]);
-        }
-
-        return proofs;
-    }
-
-    public getEverythingWithProofs(): Map<string, Map<string, [Buffer, string[]]>> | undefined {
-        if (!this.tree) {
-            return;
-        }
-
-        const proofs = new Map<string, Map<string, [Buffer, string[]]>>();
-        for (const [address] of this.values.entries()) {
-            const map = this.getValuesWithProofs(address);
-
-            proofs.set(address, map);
-        }
-
-        return proofs;
-    }
-
-    public getValues(): [Buffer, Buffer][] {
-        const entries: [Buffer, Buffer][] = [];
-
-        for (const [_address, map] of this.values.entries()) {
-            for (const [key, value] of map.entries()) {
-                const keyBuf = Buffer.from(key, 'hex');
-
-                entries.push([keyBuf, value]);
-            }
-        }
-
-        return entries;
-    }
-
-    protected getDummyValues(): Map<string, Map<string, Buffer>> {
-        const dummyValues = new Map<string, Map<string, Buffer>>();
-        const dummyMap = new Map<string, Buffer>();
-
-        // Ensure minimum tree requirements
-        dummyMap.set(MAX_HASH, Buffer.from([1]));
-        dummyMap.set(MAX_MINUS_ONE, Buffer.from([1]));
-
-        // Add dummy values for the contract
-        dummyValues.set(MAX_MINUS_ONE, dummyMap);
-
-        return dummyValues;
-    }
-
-    private ensureAddress(address: Address): void {
-        if (!this.values.has(address)) {
-            this.values.set(address, new Map());
-        }
-    }
-}
+import test from 'ava';
+import { randomAddress } from './generator.js';
+import { ReceiptMerkleTreeOld } from './old/ReceiptMerkleTree.js';
+import { ReceiptMerkleTreeNew } from './new/ReceiptMerkleTreeNew.js';
 
 test('Test ReceiptMerkleTree compatibility', (t) => {
-    const merkleOld = new ReceiptMerkleTreeOld()
-    merkleOld.updateValue("abcd1", "abc1", new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]))
-    merkleOld.updateValue("abcd1", "abc2", new Uint8Array([2, 3, 4, 5, 6, 7, 8, 9]))
-    merkleOld.updateValue("abcd1", "abc3", new Uint8Array([4, 5, 6, 7, 8, 9, 10, 11]))
-    merkleOld.updateValue("abcd1", "abc4", new Uint8Array([5, 6, 7, 8, 9, 10, 11, 12]))
-    merkleOld.updateValue("abcd1", "abc5", new Uint8Array([6, 7, 8, 9, 10, 11, 12, 13]))
-    merkleOld.updateValue("abcd1", "abc6", new Uint8Array([7, 8, 9, 10, 11, 12, 13, 14]))
-    merkleOld.updateValue("abcd1", "abc7", new Uint8Array([8, 9, 10, 11, 12, 13, 14, 15]))
-    merkleOld.updateValue("abcd1", "abc8", new Uint8Array([9, 10, 11, 12, 13, 14, 15, 16]))
-    merkleOld.generateTree()
-    const proofOld = merkleOld.getProofs()
+    const contract = randomAddress();
+    const transactionId1 = '0xaaa0';
+    const transactionId2 = '0xaaab';
+    const transactionId3 = '0xaaac';
+    const transactionId4 = '0xaaad';
+    const transactionId5 = '0xaaae';
+    const transactionId6 = '0xaaaf';
+    const transactionId7 = '0xaaa1';
+    const transactionId8 = '0xaaa2';
 
+    const merkleOld = new ReceiptMerkleTreeOld();
+    const merkleNew = new ReceiptMerkleTreeNew();
 
-    const merkleNew = new ReceiptMerkleTreeNew()
-    merkleNew.updateValue("abcd1", "abc1", new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]))
-    merkleNew.updateValue("abcd1", "abc2", new Uint8Array([2, 3, 4, 5, 6, 7, 8, 9]))
-    merkleNew.updateValue("abcd1", "abc3", new Uint8Array([4, 5, 6, 7, 8, 9, 10, 11]))
-    merkleNew.updateValue("abcd1", "abc4", new Uint8Array([5, 6, 7, 8, 9, 10, 11, 12]))
-    merkleNew.updateValue("abcd1", "abc5", new Uint8Array([6, 7, 8, 9, 10, 11, 12, 13]))
-    merkleNew.updateValue("abcd1", "abc6", new Uint8Array([7, 8, 9, 10, 11, 12, 13, 14]))
-    merkleNew.updateValue("abcd1", "abc7", new Uint8Array([8, 9, 10, 11, 12, 13, 14, 15]))
-    merkleNew.updateValue("abcd1", "abc8", new Uint8Array([9, 10, 11, 12, 13, 14, 15, 16]))
-    merkleNew.generateTree()
-    const proofNew = merkleNew.getProofs()
+    merkleOld.updateValue(contract, transactionId1, new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]));
+    merkleOld.updateValue(contract, transactionId2, new Uint8Array([2, 3, 4, 5, 6, 7, 8, 9]));
+    merkleOld.updateValue(contract, transactionId3, new Uint8Array([4, 5, 6, 7, 8, 9, 10, 11]));
+    merkleOld.updateValue(contract, transactionId4, new Uint8Array([5, 6, 7, 8, 9, 10, 11, 12]));
+    merkleOld.updateValue(contract, transactionId5, new Uint8Array([6, 7, 8, 9, 10, 11, 12, 13]));
+    merkleOld.updateValue(contract, transactionId6, new Uint8Array([7, 8, 9, 10, 11, 12, 13, 14]));
+    merkleOld.updateValue(contract, transactionId7, new Uint8Array([8, 9, 10, 11, 12, 13, 14, 15]));
+    merkleOld.updateValue(
+        contract,
+        transactionId8,
+        new Uint8Array([9, 10, 11, 12, 13, 14, 15, 16]),
+    );
 
-    t.deepEqual(merkleOld!.values, merkleNew!.values)
-    t.deepEqual(proofOld, proofNew)
-})
+    merkleNew.updateValue(contract, transactionId1, new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]));
+    merkleNew.updateValue(contract, transactionId2, new Uint8Array([2, 3, 4, 5, 6, 7, 8, 9]));
+    merkleNew.updateValue(contract, transactionId3, new Uint8Array([4, 5, 6, 7, 8, 9, 10, 11]));
+    merkleNew.updateValue(contract, transactionId4, new Uint8Array([5, 6, 7, 8, 9, 10, 11, 12]));
+    merkleNew.updateValue(contract, transactionId5, new Uint8Array([6, 7, 8, 9, 10, 11, 12, 13]));
+    merkleNew.updateValue(contract, transactionId6, new Uint8Array([7, 8, 9, 10, 11, 12, 13, 14]));
+    merkleNew.updateValue(contract, transactionId7, new Uint8Array([8, 9, 10, 11, 12, 13, 14, 15]));
+    merkleNew.updateValue(
+        contract,
+        transactionId8,
+        new Uint8Array([9, 10, 11, 12, 13, 14, 15, 16]),
+    );
+
+    merkleOld.generateTree();
+    merkleNew.generateTree();
+
+    const proofOld = merkleOld.getProofs();
+    const proofNew = merkleNew.getProofs();
+
+    t.deepEqual(merkleOld!.values, merkleNew!.values);
+    t.deepEqual(proofOld, proofNew);
+});
