+++ __test__/new/StateMerkleTreeNew.ts
@@ -1,15 +1,21 @@
-import { BufferHelper, MemorySlotData, MemorySlotPointer } from '@btc-vision/transaction';
+import { MerkleTreeNew } from './MerkleTree.js';
+import {
+    Address,
+    AddressMap,
+    BufferHelper,
+    MemorySlotData,
+    MemorySlotPointer,
+} from '@btc-vision/transaction';
 import crypto from 'crypto';
-import { MerkleTree } from './MerkleTree.js';
 
-export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotData<bigint>> {
+export class StateMerkleTreeNew extends MerkleTreeNew<MemorySlotPointer, MemorySlotData<bigint>> {
     public static TREE_TYPE: [string, string] = ['bytes32', 'bytes32'];
 
     constructor() {
-        super(StateMerkleTree.TREE_TYPE);
+        super(StateMerkleTreeNew.TREE_TYPE);
     }
 
-    public static encodePointerBuffer(contract: string, pointer: Uint8Array | Buffer): Buffer {
+    public static encodePointerBuffer(contract: Address, pointer: Uint8Array | Buffer): Buffer {
         const hash = crypto.createHash('sha256');
         hash.update(contract);
         hash.update(pointer);
@@ -17,20 +23,18 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         return hash.digest();
     }
 
-    public getProofs(): Map<string, Map<MemorySlotPointer, string[]>> {
+    public getProofs(): AddressMap<Map<MemorySlotPointer, string[]>> {
         if (!this.tree) {
             throw new Error('Merkle tree not generated');
         }
 
-        this.validate();
-
-        const proofs = new Map<string, Map<MemorySlotPointer, string[]>>();
-        for (const [address, val] of this.values.entries()) {
+        const proofs = new AddressMap<Map<MemorySlotPointer, string[]>>();
+        for (const [address, val] of this.values) {
             for (const [key, value] of val.entries()) {
                 const pointer = this.encodePointer(address, key);
                 const valueAsBuffer = Buffer.from(BufferHelper.valueToUint8Array(value));
 
-                const proof: string[] = this.tree.getProof([pointer, valueAsBuffer]);
+                const proof: string[] = this.getProofHashes([pointer, valueAsBuffer]);
                 if (!proof || !proof.length) {
                     throw new Error(`Proof not found for ${pointer.toString('hex')}`);
                 }
@@ -51,7 +55,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
 
     /** We have to replace the value of the given address and key with the new value */
     public updateValues(
-        address: string,
+        address: Address,
         val: Map<MemorySlotPointer, MemorySlotData<bigint>>,
     ): void {
         this.ensureAddress(address);
@@ -75,7 +79,11 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         this.valueChanged = valueChanged;
     }
 
-    public updateValue(address: string, key: MemorySlotPointer, val: MemorySlotData<bigint>): void {
+    public updateValue(
+        address: Address,
+        key: MemorySlotPointer,
+        val: MemorySlotData<bigint>,
+    ): void {
         if (this.frozen) {
             throw new Error('Merkle tree is frozen, cannot update value');
         }
@@ -96,7 +104,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         this.valueChanged = true;
     }
 
-    public getValue(address: string, key: MemorySlotPointer): MemorySlotData<bigint> | undefined {
+    public getValue(address: Address, key: MemorySlotPointer): MemorySlotData<bigint> | undefined {
         if (!this.values.has(address)) {
             return;
         }
@@ -110,7 +118,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
     }
 
     public getValueWithProofs(
-        address: string,
+        address: Address,
         key: MemorySlotPointer,
     ): [Uint8Array, string[]] | undefined {
         const value = this.getValue(address, key);
@@ -126,9 +134,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
             return [uint8Array, []];
         }
 
-        this.validate();
-
-        const proof: string[] = this.tree.getProof([pointer, valueAsBuffer]);
+        const proof: string[] = this.getProofHashes([pointer, valueAsBuffer]);
         if (!proof || !proof.length) {
             throw new Error(`Proof not found for ${pointer.toString('hex')}`);
         }
@@ -137,13 +143,11 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
     }
 
     public getValuesWithProofs(
-        address: string,
+        address: Address,
     ): Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]> {
         if (!this.tree) {
             throw new Error('Merkle tree not generated');
         }
-
-        this.validate();
 
         const proofs = new Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]>();
         if (!this.values.has(address)) {
@@ -159,7 +163,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
             const pointer = this.encodePointer(address, key);
             const valueAsBuffer = Buffer.from(BufferHelper.valueToUint8Array(value));
 
-            const proof: string[] = this.tree.getProof([pointer, valueAsBuffer]);
+            const proof: string[] = this.getProofHashes([pointer, valueAsBuffer]);
 
             if (!proof || !proof.length) {
                 throw new Error(`Proof not found for pointer ${pointer.toString('hex')}`);
@@ -172,20 +176,14 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
     }
 
     public getEverythingWithProofs():
-        | Map<string, Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]>>
+        | AddressMap<Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]>>
         | undefined {
         if (!this.tree) {
             return;
         }
 
-        this.validate();
-
-        const proofs = new Map<
-            string,
-            Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]>
-        >();
-
-        for (const [address] of this.values.entries()) {
+        const proofs = new AddressMap<Map<MemorySlotPointer, [MemorySlotData<bigint>, string[]]>>();
+        for (const address of this.values.keys()) {
             const map = this.getValuesWithProofs(address);
 
             proofs.set(address, map);
@@ -194,8 +192,8 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         return proofs;
     }
 
-    public encodePointer(contract: string, pointer: bigint): Buffer {
-        return StateMerkleTree.encodePointerBuffer(
+    public encodePointer(contract: Address, pointer: bigint): Buffer {
+        return StateMerkleTreeNew.encodePointerBuffer(
             contract,
             BufferHelper.pointerToUint8Array(pointer),
         );
@@ -204,7 +202,7 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
     public getValues(): [Buffer, Buffer][] {
         const entries: [Buffer, Buffer][] = [];
 
-        for (const [address, map] of this.values.entries()) {
+        for (const [address, map] of this.values) {
             for (const [key, value] of map.entries()) {
                 const pointer = this.encodePointer(address, key);
                 const valueAsBuffer = Buffer.from(BufferHelper.valueToUint8Array(value));
@@ -216,8 +214,8 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         return entries;
     }
 
-    protected getDummyValues(): Map<string, Map<MemorySlotPointer, MemorySlotData<bigint>>> {
-        const dummyValues = new Map<string, Map<MemorySlotPointer, MemorySlotData<bigint>>>();
+    protected getDummyValues(): AddressMap<Map<MemorySlotPointer, MemorySlotData<bigint>>> {
+        const dummyValues = new AddressMap<Map<MemorySlotPointer, MemorySlotData<bigint>>>();
         const dummyMap = new Map<MemorySlotPointer, MemorySlotData<bigint>>();
 
         // Ensure minimum tree requirements
@@ -225,12 +223,12 @@ export class StateMerkleTree extends MerkleTree<MemorySlotPointer, MemorySlotDat
         dummyMap.set(2n, 2n);
 
         // Add dummy values for the contract
-        dummyValues.set(MerkleTree.DUMMY_ADDRESS_NON_EXISTENT, dummyMap);
+        dummyValues.set(this.DUMMY_ADDRESS_NON_EXISTENT, dummyMap);
 
         return dummyValues;
     }
 
-    private ensureAddress(address: string): void {
+    private ensureAddress(address: Address): void {
         if (!this.values.has(address)) {
             this.values.set(address, new Map());
         }
