+++ __test__/types/ZeroValue.ts
@@ -1,3 +1,5 @@
+import { Address } from '@btc-vision/transaction';
+
 export const ZERO_HASH: string =
     '0x0000000000000000000000000000000000000000000000000000000000000000';
 
@@ -6,4 +8,4 @@ export const MAX_HASH: string =
 export const MAX_MINUS_ONE: string =
     '0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe';
 
-export const BTC_FAKE_ADDRESS: string = `fake`;
+export const BTC_FAKE_ADDRESS: Address = Address.dead();
