--- example/types/IBlockHeaderDocument.ts
@@ -1,1 +1,0 @@
-export type BlockHeaderChecksumProof = Array<[number, string[]]>;
